chore: solve few Circular Dependencies in components
