+++ packages/components/src/components/BulletList/BulletList.tsx
@@ -1,9 +1,8 @@
-import { createContext, useContext } from 'react';
-
 import styled from 'styled-components';
 
 import { SpacingValuesNew } from '@trezor/theme';
 
+import { BulletListContext } from './BulletListContext';
 import { BulletListItem } from './BulletListItem';
 import { BulletLineWidth, BulletListDirection, BulletSize } from './types';
 import {
@@ -48,26 +47,6 @@ export type BulletListProps = AllowedFrameProps & {
     'data-testid'?: string;
 };
 
-type BulletListContextValue = {
-    itemGap: SpacingValuesNew;
-    titleGap: SpacingValuesNew;
-    bulletGap: SpacingValuesNew;
-    bulletSize: BulletSize;
-    lineWidth: BulletLineWidth;
-    isOrdered: boolean;
-    direction: BulletListDirection;
-};
-
-const BulletListContext = createContext<BulletListContextValue>({
-    itemGap: 32,
-    titleGap: 8,
-    bulletGap: 24,
-    bulletSize: 'large',
-    isOrdered: false,
-    lineWidth: 2,
-    direction: 'vertical',
-});
-
 export const BulletList = ({
     gap = 32,
     bulletGap = 24,
@@ -100,7 +79,5 @@ export const BulletList = ({
         </BulletListContext.Provider>
     );
 };
-
-export const useBulletList = () => useContext(BulletListContext);
 
 BulletList.Item = BulletListItem;
