+++ packages/components/src/components/BulletList/BulletListContext.ts
@@ -1,0 +1,27 @@
+import { createContext, useContext } from 'react';
+
+import { SpacingValuesNew } from '@trezor/theme';
+
+import { BulletLineWidth, BulletListDirection, BulletSize } from './types';
+
+type BulletListContextValue = {
+    itemGap: SpacingValuesNew;
+    titleGap: SpacingValuesNew;
+    bulletGap: SpacingValuesNew;
+    bulletSize: BulletSize;
+    lineWidth: BulletLineWidth;
+    isOrdered: boolean;
+    direction: BulletListDirection;
+};
+
+export const BulletListContext = createContext<BulletListContextValue>({
+    itemGap: 32,
+    titleGap: 8,
+    bulletGap: 24,
+    bulletSize: 'large',
+    isOrdered: false,
+    lineWidth: 2,
+    direction: 'vertical',
+});
+
+export const useBulletList = () => useContext(BulletListContext);
