+++ packages/components/src/components/Collapsible/CollapsibleContext.tsx
@@ -1,0 +1,19 @@
+import { createContext, useContext } from 'react';
+
+import { SpacingValues, spacings } from '@trezor/theme';
+
+type CollapsibleContextProps = {
+    isOpen: boolean;
+    toggle: (isOpen: boolean) => void;
+    gap?: SpacingValues;
+    contentId: string;
+};
+
+export const CollapsibleContext = createContext<CollapsibleContextProps>({
+    isOpen: false,
+    toggle: () => {},
+    gap: spacings.zero,
+    contentId: '',
+});
+
+export const useCollapsible = () => useContext(CollapsibleContext);
