+++ packages/components/src/components/List/List.tsx
@@ -1,11 +1,12 @@
-import { createContext, useContext } from 'react';
+import { ReactNode } from 'react';
 
 import styled from 'styled-components';
 
 import { SpacingValues, spacings, spacingsPx } from '@trezor/theme';
 
+import { ListContext } from './ListContext';
 import { ListItem } from './ListItem';
-import { uiAlignments } from '../../config/types';
+import { BulletVerticalAlignment, ListStyleType, ListVariant } from './types';
 import {
     FrameProps,
     FramePropsKeys,
@@ -13,7 +14,7 @@ import {
     withFrameProps,
 } from '../../utils/frameProps';
 import { TransientProps, makePropsTransient } from '../../utils/transientProps';
-import { Text, textVariants } from '../typography/Text/Text';
+import { Text } from '../typography/Text/Text';
 import {
     TextProps,
     TextPropsKeys,
@@ -21,35 +22,18 @@ import {
     withTextProps,
 } from '../typography/utils';
 
-type ListStyleType =
-    | 'disc'
-    | 'circle'
-    | 'square'
-    | 'decimal'
-    | 'decimal-leading-zero'
-    | 'lower-roman'
-    | 'upper-roman'
-    | 'lower-alpha'
-    | 'upper-alpha';
-
 export const allowedListFrameProps = [
     'margin',
     'width',
     'overflow',
 ] as const satisfies FramePropsKeys[];
-type AllowedFrameProps = Pick<FrameProps, (typeof allowedListFrameProps)[number]>;
+export type AllowedFrameProps = Pick<FrameProps, (typeof allowedListFrameProps)[number]>;
 
 export const allowedListTextProps = [
     'typographyStyle',
     'textWrap',
 ] as const satisfies TextPropsKeys[];
-type AllowedTextProps = Pick<TextProps, (typeof allowedListTextProps)[number]>;
-
-export const listVariants = textVariants;
-export type ListVariant = (typeof listVariants)[number];
-
-export const bulletVerticalAlignments = uiAlignments;
-export type BulletVerticalAlignment = (typeof bulletVerticalAlignments)[number];
+export type AllowedTextProps = Pick<TextProps, (typeof allowedListTextProps)[number]>;
 
 type ContainerProps = TransientProps<AllowedFrameProps & AllowedTextProps> & {
     $gap: SpacingValues;
@@ -71,26 +55,13 @@ const Container = styled.ul<ContainerProps>`
 export type ListProps = AllowedFrameProps &
     AllowedTextProps & {
         gap?: SpacingValues;
-        children: React.ReactNode;
-        bulletComponent?: React.ReactNode;
+        children: ReactNode;
+        bulletComponent?: ReactNode;
         bulletGap?: SpacingValues;
         bulletAlignment?: BulletVerticalAlignment;
         variant?: ListVariant;
         listStyleType?: ListStyleType;
     };
-
-type ListContextValue = {
-    bulletGap: SpacingValues;
-    bulletAlignment: BulletVerticalAlignment;
-    bulletComponent: React.ReactNode;
-    listStyleType?: ListStyleType;
-};
-
-const ListContext = createContext<ListContextValue>({
-    bulletGap: spacings.md,
-    bulletAlignment: 'center',
-    bulletComponent: null as React.ReactNode,
-});
 
 export const List = ({
     gap = spacings.xs,
@@ -121,7 +92,5 @@ export const List = ({
         </ListContext.Provider>
     );
 };
-
-export const useList = () => useContext(ListContext);
 
 List.Item = ListItem;
