+++ packages/components/src/components/List/ListContext.tsx
@@ -1,0 +1,20 @@
+import { ReactNode, createContext, useContext } from 'react';
+
+import { SpacingValues, spacings } from '@trezor/theme';
+
+import { BulletVerticalAlignment, ListStyleType } from './types';
+
+type ListContextValue = {
+    bulletGap: SpacingValues;
+    bulletAlignment: BulletVerticalAlignment;
+    bulletComponent: ReactNode;
+    listStyleType?: ListStyleType;
+};
+
+export const ListContext = createContext<ListContextValue>({
+    bulletGap: spacings.md,
+    bulletAlignment: 'center',
+    bulletComponent: null as ReactNode,
+});
+
+export const useList = () => useContext(ListContext);
